backend/requirements.txt
fastapi
uvicorn
python-dotenv
groq
deepgram-sdk
edge-tts
pypdf
python-multipart
websockets
requests